refactor(bloat): split self-heal commit-path out of gitattributes_union
Adding the CURATED_DOC_UNION_PATHS category pushed gitattributes_union.py over the 300-LOC bloat limit. Split the git self-heal commit-path (self_heal_gitattributes, HealResult, atomic-write/rollback helpers) into lib/gitattributes_selfheal.py, mirroring the existing test split. gitattributes_union.py is now 141 LOC (pure merge-logic SSoT, stdlib-only top-level so the adopt file-path loader stays happy); gitattributes_selfheal.py is 191 — both <=300, no crossing. Callers (setup_iterate_worktree, the selfheal test) updated. 3270 shared tests green.

Run-ID: iterate-2026-06-12-union-curated-agent-docs
